requirements.txt
streamlit
pandas
numpy
plotly
seaborn
matplotlib
openpyxl
openai
scikit-learn
networkx
python-telegram-bot